Added bus monitor on nodeID 0xfe. The bus monitor can be accessed by setting a nodeID to 0xfe and calling the obj.read() command. Unlike the typical obj.read() which returns a read success or failure flag, the value returned by a bus monitor node will be the nodeID of the node sending a message. All failures to read a message by the bus monitor will return 0xff, which is also now an invalid nodeID. This ensures that when a bus monitor successfully reads date, the bus monitor will know where which node that data originated from.
We decided to incoroporate the bus monitor as a special type of slave node. This allows us to leave the obj.write() and obj.read() commands superficially unchanged with no additional inputs or outputs added to them. Their innerworkings have been modified to handle the additional bus monitor case. We hope this will leave the code easy to execute and understand for the average user. We expect the bus Monitor function will mostly be used by more advanced users who will be adept and handling the small changes in obj.read() output flag as described above.

Restricted nodeID 0xff to reserved. Setting a nodeID of 0xff will cause an error

Changed the wording of debug display code:
Fail - indiciates a function call that was cancelled or terminated
Status - indicates more detail on what a function is doing
Warn - indicates that some operation occurred but was compensated for
Note - describes ways to fix warnings and failures
fcall - lists the function called by the library